+++ code/packages/haskell/zip/test/Spec.hs
@@ -1,5 +1,8 @@
 import Test.Hspec
+import qualified PortableConformanceSpec
 import ZipSpec
 
 main :: IO ()
-main = hspec spec
+main = hspec $ do
+    spec
+    PortableConformanceSpec.spec